remove instances of VectorObtainer
