+++ usecases/modules/modules.go
@@ -23,8 +23,6 @@ import (
 	"github.com/semi-technologies/weaviate/entities/moduletools"
 	"github.com/semi-technologies/weaviate/entities/schema"
 	"github.com/semi-technologies/weaviate/entities/search"
-	"github.com/semi-technologies/weaviate/entities/vectorindex/hnsw"
-	"github.com/semi-technologies/weaviate/usecases/config"
 	"github.com/sirupsen/logrus"
 )
 
@@ -716,110 +714,4 @@ func (m *Provider) BackupBackend(backend string) (modulecapabilities.BackupBacke
 		}
 	}
 	return nil, errors.Errorf("backup: %s not found", backend)
-}
-
-func (m *Provider) UpdateVector(ctx context.Context, object *models.Object,
-	repo modulecapabilities.VectorRepo, logger logrus.FieldLogger,
-) error {
-	class, err := m.getClass(object.Class)
-	if err != nil {
-		return err
-	}
-
-	vectorizerName, idxCfg, err := m.getClassVectorizer(object.Class)
-	if err != nil {
-		return err
-	}
-
-	hnswConfig, ok := idxCfg.(hnsw.UserConfig)
-	if !ok {
-		return fmt.Errorf("vector index config (%T) is not of type HNSW, "+
-			"but objects manager is restricted to HNSW", idxCfg)
-	}
-
-	if vectorizerName == config.VectorizerModuleNone {
-		err = validateVectorPresent(object, hnswConfig, logger)
-		if err != nil {
-			return err
-		}
-
-		return nil
-	}
-
-	if hnswConfig.Skip {
-		logger.WithField("className", object.Class).
-			WithField("vectorizer", vectorizerName).
-			Warningf("this class is configured to skip vector indexing, "+
-				"but a vector was generated by the %q vectorizer. "+
-				"This vector will be ignored. If you meant to index "+
-				"the vector, make sure to set vectorIndexConfig.skip to 'false'. If the previous "+
-				"setting is correct, make sure you set vectorizer to 'none' in the schema and "+
-				"provide a null-vector (i.e. no vector) at import time.", vectorizerName)
-	}
-
-	var found modulecapabilities.Module
-	for modName := range class.ModuleConfig.(map[string]interface{}) {
-		if err := m.ValidateVectorizer(modName); err == nil {
-			found = m.GetByName(modName)
-			break
-		}
-	}
-
-	if found == nil {
-		return fmt.Errorf(
-			"no vectorizer found for class %q: %w", object.Class, err)
-	}
-
-	cfg := NewClassBasedModuleConfig(class, found.Name())
-
-	if vectorizer, ok := found.(modulecapabilities.Vectorizer); ok {
-		if err := vectorizer.VectorizeObject(ctx, object, cfg); err != nil {
-			return fmt.Errorf("update vector: %w", err)
-		}
-	} else {
-		refVectorizer := found.(modulecapabilities.ReferenceVectorizer)
-		if err := refVectorizer.VectorizeObject(
-			ctx, object, cfg, repo.ReferenceVectorSearch); err != nil {
-			return fmt.Errorf("update vector: %w", err)
-		}
-	}
-
-	return nil
-}
-
-func (m *Provider) VectorizerName(className string) (string, error) {
-	name, _, err := m.getClassVectorizer(className)
-	if err != nil {
-		return "", err
-	}
-	return name, nil
-}
-
-func validateVectorPresent(obj *models.Object,
-	hnswConfig hnsw.UserConfig, logger logrus.FieldLogger,
-) error {
-	if hnswConfig.Skip && len(obj.Vector) > 0 {
-		logger.WithField("className", obj.Class).
-			Warningf("this class is configured to skip vector indexing, " +
-				"but a vector was explicitly provided. " +
-				"This vector will be ignored. If you meant to index " +
-				"the vector, make sure to set vectorIndexConfig.skip to 'false'. If the previous " +
-				"setting is correct, make sure you set vectorizer to 'none' in the schema and " +
-				"provide a null-vector (i.e. no vector) at import time.")
-	}
-
-	return nil
-}
-
-func (m *Provider) getClassVectorizer(className string) (string, interface{}, error) {
-	sch := m.schemaGetter.GetSchemaSkipAuth()
-
-	class := sch.FindClassByName(schema.ClassName(className))
-	if class == nil {
-		// this should be impossible by the time this method gets called, but let's
-		// be 100% certain
-		return "", nil, fmt.Errorf("class %s not present", className)
-	}
-
-	return class.Vectorizer, class.VectorIndexConfig, nil
 }
