+++ usecases/modules/vectorizer.go
@@ -13,12 +13,34 @@ package modules
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/pkg/errors"
 	"github.com/semi-technologies/weaviate/entities/models"
 	"github.com/semi-technologies/weaviate/entities/modulecapabilities"
 	"github.com/semi-technologies/weaviate/entities/schema"
-	"github.com/semi-technologies/weaviate/usecases/objects"
+	"github.com/semi-technologies/weaviate/entities/vectorindex/hnsw"
+	"github.com/semi-technologies/weaviate/usecases/config"
+	"github.com/sirupsen/logrus"
+)
+
+const (
+	errorVectorizerCapability = "module %q exists, but does not provide the " +
+		"Vectorizer or ReferenceVectorizer capability"
+
+	errorVectorIndexType = "vector index config (%T) is not of type HNSW, " +
+		"but objects manager is restricted to HNSW"
+
+	warningVectorIgnored = "This vector will be ignored. If you meant to index " +
+		"the vector, make sure to set vectorIndexConfig.skip to 'false'. If the previous " +
+		"setting is correct, make sure you set vectorizer to 'none' in the schema and " +
+		"provide a null-vector (i.e. no vector) at import time."
+
+	warningSkipVectorGenerated = "this class is configured to skip vector indexing, " +
+		"but a vector was generated by the %q vectorizer. " + warningVectorIgnored
+
+	warningSkipVectorProvided = "this class is configured to skip vector indexing, " +
+		"but a vector was explicitly provided. " + warningVectorIgnored
 )
 
 func (m *Provider) ValidateVectorizer(moduleName string) error {
@@ -30,33 +52,10 @@ func (m *Provider) ValidateVectorizer(moduleName string) error {
 	_, okVec := mod.(modulecapabilities.Vectorizer)
 	_, okRefVec := mod.(modulecapabilities.ReferenceVectorizer)
 	if !okVec && !okRefVec {
-		return errors.Errorf("module %q exists, but does not provide the "+
-			"Vectorizer or ReferenceVectorizer capability", moduleName)
+		return errors.Errorf(errorVectorizerCapability, moduleName)
 	}
 
 	return nil
-}
-
-func (m *Provider) Vectorizer(moduleName, className string) (objects.Vectorizer, error) {
-	mod := m.GetByName(moduleName)
-	if mod == nil {
-		return nil, errors.Errorf("no module with name %q present", moduleName)
-	}
-
-	vec, ok := mod.(modulecapabilities.Vectorizer)
-	if !ok {
-		return nil, errors.Errorf("module %q exists, but does not provide the "+
-			"Vectorizer capability", moduleName)
-	}
-
-	sch := m.schemaGetter.GetSchemaSkipAuth()
-	class := sch.FindClassByName(schema.ClassName(className))
-	if class == nil {
-		return nil, errors.Errorf("class %q not found in schema", className)
-	}
-
-	cfg := NewClassBasedModuleConfig(class, moduleName)
-	return NewObjectsVectorizer(vec, cfg), nil
 }
 
 func (m *Provider) UsingRef2Vec(className string) bool {
@@ -80,36 +79,86 @@ func (m *Provider) UsingRef2Vec(className string) bool {
 	return false
 }
 
-type ObjectsVectorizer struct {
-	modVectorizer modulecapabilities.Vectorizer
-	cfg           *ClassBasedModuleConfig
-}
-
-func NewObjectsVectorizer(vec modulecapabilities.Vectorizer,
-	cfg *ClassBasedModuleConfig,
-) *ObjectsVectorizer {
-	return &ObjectsVectorizer{modVectorizer: vec, cfg: cfg}
-}
-
-func (ov *ObjectsVectorizer) UpdateObject(ctx context.Context,
-	obj *models.Object,
+func (m *Provider) UpdateVector(ctx context.Context, object *models.Object,
+	repo modulecapabilities.VectorRepo, logger logrus.FieldLogger,
 ) error {
-	return ov.modVectorizer.VectorizeObject(ctx, obj, ov.cfg)
+	class, err := m.getClass(object.Class)
+	if err != nil {
+		return err
+	}
+
+	vectorizerName, idxCfg, err := m.getClassVectorizer(object.Class)
+	if err != nil {
+		return err
+	}
+
+	hnswConfig, ok := idxCfg.(hnsw.UserConfig)
+	if !ok {
+		return fmt.Errorf(errorVectorIndexType, idxCfg)
+	}
+
+	if vectorizerName == config.VectorizerModuleNone {
+		if hnswConfig.Skip && len(object.Vector) > 0 {
+			logger.WithField("className", object.Class).
+				Warningf(warningSkipVectorProvided)
+		}
+
+		return nil
+	}
+
+	if hnswConfig.Skip {
+		logger.WithField("className", object.Class).
+			WithField("vectorizer", vectorizerName).
+			Warningf(warningSkipVectorGenerated, vectorizerName)
+	}
+
+	var found modulecapabilities.Module
+	for modName := range class.ModuleConfig.(map[string]interface{}) {
+		if err := m.ValidateVectorizer(modName); err == nil {
+			found = m.GetByName(modName)
+			break
+		}
+	}
+
+	if found == nil {
+		return fmt.Errorf(
+			"no vectorizer found for class %q: %w", object.Class, err)
+	}
+
+	cfg := NewClassBasedModuleConfig(class, found.Name())
+
+	if vectorizer, ok := found.(modulecapabilities.Vectorizer); ok {
+		if err := vectorizer.VectorizeObject(ctx, object, cfg); err != nil {
+			return fmt.Errorf("update vector: %w", err)
+		}
+	} else {
+		refVectorizer := found.(modulecapabilities.ReferenceVectorizer)
+		if err := refVectorizer.VectorizeObject(
+			ctx, object, cfg, repo.ReferenceVectorSearch); err != nil {
+			return fmt.Errorf("update reference vector: %w", err)
+		}
+	}
+
+	return nil
 }
 
-type ObjectsReferenceVectorizer struct {
-	modVectorizer modulecapabilities.ReferenceVectorizer
-	cfg           *ClassBasedModuleConfig
+func (m *Provider) VectorizerName(className string) (string, error) {
+	name, _, err := m.getClassVectorizer(className)
+	if err != nil {
+		return "", err
+	}
+	return name, nil
 }
 
-func NewObjectsReferenceVectorizer(vec modulecapabilities.ReferenceVectorizer,
-	cfg *ClassBasedModuleConfig,
-) *ObjectsReferenceVectorizer {
-	return &ObjectsReferenceVectorizer{modVectorizer: vec, cfg: cfg}
-}
+func (m *Provider) getClassVectorizer(className string) (string, interface{}, error) {
+	sch := m.schemaGetter.GetSchemaSkipAuth()
 
-func (ov *ObjectsReferenceVectorizer) UpdateObject(ctx context.Context,
-	obj *models.Object, findRefVecsFn modulecapabilities.FindRefVectorsFn,
-) error {
-	return ov.modVectorizer.VectorizeObject(ctx, obj, ov.cfg, findRefVecsFn)
+	class := sch.FindClassByName(schema.ClassName(className))
+	if class == nil {
+		// this should be impossible by the time this method gets called, but let's
+		// be 100% certain
+		return "", nil, fmt.Errorf("class %s not present", className)
+	}
+
+	return class.Vectorizer, class.VectorIndexConfig, nil
 }
