+++ usecases/modules/vectorizer_test.go
@@ -12,7 +12,6 @@
 package modules
 
 import (
-	"context"
 	"fmt"
 	"testing"
 
@@ -20,57 +19,7 @@ import (
 	"github.com/semi-technologies/weaviate/entities/modulecapabilities"
 	"github.com/semi-technologies/weaviate/entities/schema"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
-
-func TestProvider_Vectorizer(t *testing.T) {
-	t.Run("when there are no models registered", func(t *testing.T) {
-		p := NewProvider()
-		_, err := p.Vectorizer("some-module", "MyClass")
-		require.NotNil(t, err)
-		assert.Equal(t, "no module with name \"some-module\" present", err.Error())
-	})
-
-	t.Run("module exist, but doesn't provide vectorizer", func(t *testing.T) {
-		p := NewProvider()
-		p.Register(newDummyModule("some-module", ""))
-		_, err := p.Vectorizer("some-module", "MyClass")
-		require.NotNil(t, err)
-		assert.Equal(t, "module \"some-module\" exists, but does not provide the "+
-			"Vectorizer capability", err.Error())
-	})
-	t.Run("module exists, but the class doesn't", func(t *testing.T) {
-		p := NewProvider()
-		p.SetSchemaGetter(&fakeSchemaGetter{schema.Schema{}})
-		p.Register(newDummyModule("some-module", modulecapabilities.Text2Vec))
-		_, err := p.Vectorizer("some-module", "MyClass")
-		require.NotNil(t, err)
-		assert.Equal(t, "class \"MyClass\" not found in schema", err.Error())
-	})
-
-	t.Run("module exist, and provides a vectorizer", func(t *testing.T) {
-		p := NewProvider()
-		sch := schema.Schema{
-			Objects: &models.Schema{
-				Classes: []*models.Class{
-					{
-						Class: "MyClass",
-					},
-				},
-			},
-		}
-		p.SetSchemaGetter(&fakeSchemaGetter{sch})
-		p.Register(newDummyModule("some-module", modulecapabilities.Text2Vec))
-		vec, err := p.Vectorizer("some-module", "MyClass")
-		require.Nil(t, err)
-
-		obj := &models.Object{Class: "Test"}
-		err = vec.UpdateObject(context.Background(), obj)
-		require.Nil(t, err)
-
-		assert.Equal(t, models.C11yVector{1, 2, 3}, obj.Vector)
-	})
-}
 
 func TestProvider_ValidateVectorizer(t *testing.T) {
 	t.Run("with vectorizer module", func(t *testing.T) {
